requirements.txt
Flask==2.0.2
mysql-connector-python==8.0.28
Werkzeug==2.0.2
